fix(window): setup primary window on post startup
this is a mini fix for the one second white flicker when primary window
is created in the default plugin. creating it on a poststartup phase
(after the camera is setup) seems to bring the white flicker down to
milliseconds. will still need a permanent fix to completely eliminating
it.